// src/enum/app.enums.ts
export enum CookieStorageEnum {
  AccessToken = "access_token",
  RefeshToken = "refesh_token",
}

export enum LocalStorageEnum {
  UserInfo = "user_info",
}

export enum ServiceType {
  Product = 1,
  Workout = 2,
}

export enum ServiceBookingTypeEnum {
  Regular = 1,
  Bundle = 2,
}

export enum PaidStatus {
  Unpaid = 0,
  Paid = 1,
  Pending = 3,
  Refunded = 4,
}